BurgerComstuctor divided to separate components
