+++ src/components/app-header/app-header.jsx
@@ -9,12 +9,12 @@ function AppHeader() {
     <header className={`${styles.header}`}>
       <div className={`${styles.headerBox}`}>
         <div className={styles.leftIconBox}>
-          <a href='' className={`pt-4 pb-4 pl-5 pr-5 ${styles.navItem}`}>
+          <a href='#' className={`pt-4 pb-4 pl-5 pr-5 ${styles.navItem}`}>
             <BurgerIcon type='primary' />
             <p className='ml-2 text text_type_main-default'>Конструктор</p>
           </a>
 
-          <a href='' className={`pt-4 pb-4 pl-5 pr-5 ${styles.navItem}`}>
+          <a href='#' className={`pt-4 pb-4 pl-5 pr-5 ${styles.navItem}`}>
             <ListIcon type='secondary' />
             <p className='ml-2 text text_type_main-default text_color_inactive'>Лента заказов</p>
           </a>
@@ -24,7 +24,7 @@ function AppHeader() {
         </div>
 
         <div className={`pt-4 pb-4 pl-5 pr-5 ${styles.rightIconBox}`}>
-          <a href='' className={styles.navItem}>
+          <a href='#' className={styles.navItem}>
             <ProfileIcon type='secondary' />
             <p className='ml-2 text text_type_main-default text_color_inactive'>Личный кабинет</p>
           </a>
