+++ src/components/elem-bottom/elem-bottom.module.css
@@ -1,0 +1,6 @@
+.elemBottom {
+  display: flex;
+  justify-content: right;
+  align-items: center;
+  padding-right: 16px;
+}
