+++ src/components/elem-list/elem-list.module.css
@@ -1,0 +1,6 @@
+.elemList {
+  display: flex;
+  justify-content: right;
+  align-items: center;
+  padding-right: 8px;
+}
