+++ src/components/elem-top/elem-top.module.css
@@ -1,0 +1,6 @@
+.elemTop {
+  display: flex;
+  justify-content: right;
+  align-items: center;
+  padding-right: 16px;
+}
